@@ -13,28 +13,25 @@
   "private": true,
   "dependencies": {
     "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "5.2.8",
-    "@angular/cdk": "5.2.4",
-    "@angular/common": "5.2.8",
-    "@angular/compiler": "5.2.8",
-    "@angular/compiler-cli": "^5.1.0",
-    "@angular/core": "5.2.8",
+    "@angular/animations": "6.0.9",
+    "@angular/cdk": "6.3.3",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "6.0.9",
+    "@angular/compiler-cli": "6.0.9",
+    "@angular/core": "6.0.9",
     "@angular/flex-layout": "5.0.0-beta.13",
-    "@angular/forms": "^5.1.0",
-    "@angular/http": "5.2.8",
-    "@angular/material": "5.2.4",
-    "@angular/platform-browser": "5.2.8",
-    "@angular/platform-browser-dynamic": "5.2.8",
-    "@angular/platform-server": "^5.1.0",
-    "@angular/router": "5.2.8",
+    "@angular/forms": "6.0.9",
+    "@angular/http": "6.0.9",
+    "@angular/material": "6.3.3",
+    "@angular/platform-browser": "6.0.9",
+    "@angular/platform-browser-dynamic": "6.0.9",
+    "@angular/platform-server": "6.0.9",
+    "@angular/router": "6.0.9",
     "@ngrx/effects": "5.2.0",
     "@ngrx/router-store": "5.2.0",
     "@ngrx/store": "5.2.0",
     "@ngrx/store-devtools": "5.2.0",
     "@ngx-translate/core": "9.1.1",
-    "@progress/kendo-angular-dropdowns": "^1.2.0",
-    "@progress/kendo-angular-l10n": "^1.0.3",
-    "@progress/kendo-theme-default": "^2.41.2",
     "@types/html2canvas": "^0.5.35",
     "@types/jasmine": "^2.5.54",
     "amexio-ng-extensions": "^3.2.0",
@@ -58,18 +55,20 @@
     "ngx-cookie-service": "1.0.10",
     "perfect-scrollbar": "1.3.0",
     "primeng": "^5.2.0-rc.2",
-    "prismjs": "1.11.0",
-    "rxjs": "^5.4.3",
+    "prismjs": "1.15.0",
+    "@types/prismjs": "1.9.0",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
     "ts-node": "^2.1.2",
     "tslint": "^4.5.1",
-    "typescript": "2.6.2",
-    "zone.js": "^0.8.17"
+    "typescript": "2.7.2",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.2",
-    "@angular/compiler-cli": "^4.2.0",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.9",
     "@types/jasmine": "^2.5.38",
-    "@types/node": "^6.0.88",
+    "@types/node": "^6.0.114",
     "codelyzer": "^2.0.1",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "^3.2.0",
@@ -82,6 +81,7 @@
     "protractor": "^5.1.2",
     "ts-node": "^2.0.0",
     "tslint": "^4.4.2",
-    "typescript": "2.6.2"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
