@@ -1,22 +1,43 @@
 import { Injectable } from '@angular/core';
 import { MatSidenav } from '@angular/material';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class FuseMatSidenavHelperService
 {
     sidenavInstances: MatSidenav[];
 
+    /**
+     * Constructor
+     */
     constructor()
     {
         this.sidenavInstances = [];
     }
 
-    setSidenav(id, instance)
+    // -----------------------------------------------------------------------------------------------------
+    // @ Accessors
+    // -----------------------------------------------------------------------------------------------------
+
+    /**
+     * Set sidenav
+     *
+     * @param id
+     * @param instance
+     */
+    setSidenav(id, instance): void
     {
         this.sidenavInstances[id] = instance;
     }
 
-    getSidenav(id)
+    /**
+     * Get sidenav
+     *
+     * @param id
+     * @returns {any}
+     */
+    getSidenav(id): any
     {
         return this.sidenavInstances[id];
     }
