@@ -1,23 +1,51 @@
 import { MediaChange, ObservableMedia } from '@angular/flex-layout';
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { BehaviorSubject } from 'rxjs';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class FuseMatchMediaService
 {
     activeMediaQuery: string;
     onMediaChange: BehaviorSubject<string> = new BehaviorSubject<string>('');
 
-    constructor(private observableMedia: ObservableMedia)
+    /**
+     * Constructor
+     *
+     * @param {ObservableMedia} _observableMedia
+     */
+    constructor(
+        private _observableMedia: ObservableMedia
+    )
     {
+        // Set the defaults
         this.activeMediaQuery = '';
 
-        this.observableMedia.subscribe((change: MediaChange) => {
-            if ( this.activeMediaQuery !== change.mqAlias )
-            {
-                this.activeMediaQuery = change.mqAlias;
-                this.onMediaChange.next(change.mqAlias);
-            }
-        });
+        // Initialize
+        this._init();
+
     }
+
+    // -----------------------------------------------------------------------------------------------------
+    // @ Private methods
+    // -----------------------------------------------------------------------------------------------------
+
+    /**
+     * Initialize
+     *
+     * @private
+     */
+    private _init(): void
+    {
+        this._observableMedia
+            .subscribe((change: MediaChange) => {
+                if ( this.activeMediaQuery !== change.mqAlias )
+                {
+                    this.activeMediaQuery = change.mqAlias;
+                    this.onMediaChange.next(change.mqAlias);
+                }
+            });
+    }
+
 }
